# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "104-Project OpenApi",
        "description": "104-Project OpenApi description",
        "contact": {
            "email": "a8732987@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-host.dev.app",
            "description": "測試區主機"
        },
        {
            "url": "https://api-host.production.app",
            "description": "正式區主機"
        },
        {
            "url": "http://localhost",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/api/profile": {
            "get": {
                "tags": [
                    "Time"
                ],
                "summary": "取得此前時間 Summary",
                "description": "取得此前時間 Description",
                "operationId": "time",
                "responses": {
                    "200": {
                        "description": "請求成功"
                    }
                }
            }
        },
        "/api/course/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "取得課程詳情",
                "description": "取得課程詳情",
                "operationId": "courseShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "資源不存在"
                    }
                }
            }
        },
        "/api/courses": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "取得所有課程詳情",
                "description": "取得所有課程詳情",
                "operationId": "AllCoursesShow",
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "找不到課程"
                    }
                },
                "security": [
                    {
                        "Authenticate": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "新增課程",
                "description": "請求時需要附上JWT驗證",
                "operationId": "courseCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "課程名稱",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "課程描述",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "outline",
                        "in": "query",
                        "description": "課程大綱",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "資源不存在"
                    }
                },
                "security": [
                    {
                        "Authenticate": []
                    }
                ]
            }
        },
        "/api/courses/{courseId}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "取得單一課程詳情",
                "description": "取得單一課程詳情",
                "operationId": "courseShow",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Course Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "找不到對應課程"
                    }
                },
                "security": [
                    {
                        "Authenticate": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "刪除文章",
                "description": "刪除文章",
                "operationId": "courseDelete",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Course Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "資源不存在"
                    }
                },
                "security": [
                    {
                        "Authenticate": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "put": {
                "tags": [
                    "Course"
                ],
                "summary": "更新課程",
                "description": "更新課程",
                "operationId": "courseUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "課程名稱",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "課程描述",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "outline",
                        "in": "query",
                        "description": "課程大綱",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "資源不存在"
                    }
                },
                "security": [
                    {
                        "Authenticate": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authenticate": {
                "type": "apiKey",
                "name": "Authenticate",
                "in": "header"
            }
        }
    }
}